requirements.txt
PySide6>=6.4.0
polib>=1.1.1
sqlalchemy>=1.4.0
Pillow>=9.0.0
